# Auto-vectorize arrays `swap`