--- a/stl/inc/utility
+++ b/stl/inc/utility
@@ -115,10 +115,7 @@ _CONSTEXPR20 void swap(_Ty (&_Left)[_Size], _Ty (&_Right)[_Size]) noexcept(_Is_n
     }
 
     if constexpr (_Is_trivially_swappable_v<_Ty>) {
-#if _HAS_CXX20
-        if (!_STD is_constant_evaluated())
-#endif // _HAS_CXX20
-        {
+        if (!_STD _Is_constant_evaluated()) {
             _STD _Swap_trivial_arrays(_Left, _Right);
             return;
         }