Resolves #2683

# 📜 The Approach

Instead of trying to call a separately compiled implementation in the import library, reimplement the specific case in headers the way that compiler is able to vectorize it.

Do `memcpy` swap by portions of nice length, and then tail, so that compiler can use registers as immediate storage, vector registers when possible. Due to compile-time known length this will unroll perfectly.

I didn't investigate a lot which length is nice, I think 64 is a good choice: 
 * Not too much stack consumed in debug mode where there will be actual stack allocation
 * Enough to use AVX-512 if available
 * Not too much to use only SSE2
 * I hope it is good for Arm64 too 🤷 

# ⚖️  Self swap check

[[utility.swap]/7](https://eel.is/c++draft/utility.swap#7) says:
 > _Effects:_ As if by `swap_ranges(a, a + N, b)`.

[[alg.swap]/2](https://eel.is/c++draft/alg.swap#2) says:
> _Preconditions:_ The two ranges [first1, last1) and [first2, last2) do not overlap[.](https://eel.is/c++draft/alg.swap#2.sentence-1)

Therefore it is UB to swap overlapping arrays.

~~Added an assert for that and removed the check.~~

@frederick-vs-ja reported LWG-4165, and the check preserved.

# ⏱️ Benchmark results

Benchmark                       |      main    |       this    
--------------------------------|--------------|---------------
std_swap<1, uint8_t>            |   1.22 ns    |    1.20 ns    
std_swap<5, uint8_t>            |   2.83 ns    |    1.44 ns    
std_swap<15, uint8_t>           |   10.7 ns    |    1.92 ns    
std_swap<26, uint8_t>           |   14.9 ns    |    1.93 ns    
std_swap<38, uint8_t>           |   25.6 ns    |    2.41 ns    
std_swap<60, uint8_t>           |   32.4 ns    |    2.64 ns    
std_swap<125, uint8_t>          |   62.0 ns    |    4.31 ns    
std_swap<800, uint8_t>          |    399 ns    |    22.0 ns    
std_swap<3000, uint8_t>         |   1490 ns    |    90.4 ns    
std_swap<9000, uint8_t>         |   4365 ns    |     203 ns    
std_swap_ranges<uint8_t>/1      |   1.97 ns    |    1.91 ns    
std_swap_ranges<uint8_t>/5      |   4.03 ns    |    3.58 ns    
std_swap_ranges<uint8_t>/15     |   5.26 ns    |    5.01 ns    
std_swap_ranges<uint8_t>/26     |   3.55 ns    |    3.12 ns    
std_swap_ranges<uint8_t>/38     |   5.34 ns    |    4.53 ns    
std_swap_ranges<uint8_t>/60     |   5.61 ns    |    5.05 ns    
std_swap_ranges<uint8_t>/125    |   7.01 ns    |    6.50 ns    
std_swap_ranges<uint8_t>/800    |   19.1 ns    |    16.9 ns    
std_swap_ranges<uint8_t>/3000   |   81.8 ns    |    79.4 ns    
std_swap_ranges<uint8_t>/9000   |    139 ns    |     138 ns    

# 🥇 Results interpretation
 * Improvement even for arrays of 5 elements!
 * Sometimes wins and sometimes loses to manually vectorized implementation
    * The auto vectorization of middle sized arrays benefits from knowing the number of elements in advance and full unrolling
    * The manual vectorization of large sized array benefits from CPU dispatch and small loop body