@philnik777 2024-09-29
Can't you just throw in a `__restrict`? Works just fine with Clang/libc++: https://godbolt.org/z/Ez1hM11jG