@AlexGuteniev 2024-09-29
> Can't you just throw in a `__restrict`? Works just fine with Clang/libc++:

No, unfortunately. Pragma loop ivdep also doesn't help