@AlexGuteniev 2024-10-18
I want to note that `ranges::swap` does not have any self-swap check, unlike `std::swap`.