@StephanTLavavej 2024-10-22
Thanks! :cat2: I pushed fairly minor changes.

Results on my 5950X:

Benchmark                       |  Before |   After | Speedup
--------------------------------|--------:|--------:|-------:
`std_swap<1, uint8_t>`          | 2.98 ns | 3.63 ns | 0.82
`std_swap<5, uint8_t>`          | 4.72 ns | 3.84 ns | 1.23
`std_swap<15, uint8_t>`         | 8.33 ns | 3.87 ns | 2.15
`std_swap<26, uint8_t>`         | 24.0 ns | 3.65 ns | 6.58
`std_swap<38, uint8_t>`         | 38.7 ns | 3.86 ns | 10.03
`std_swap<60, uint8_t>`         | 62.5 ns | 3.87 ns | 16.15
`std_swap<125, uint8_t>`        |  129 ns | 4.75 ns | 27.16
`std_swap<800, uint8_t>`        |  816 ns | 21.4 ns | 38.13
`std_swap<3000, uint8_t>`       | 3019 ns | 80.6 ns | 37.46
`std_swap<9000, uint8_t>`       | 9262 ns |  241 ns | 38.43
`std_swap_ranges<uint8_t>/1`    | 4.70 ns | 4.48 ns | 1.05
`std_swap_ranges<uint8_t>/5`    | 5.60 ns | 6.22 ns | 0.90
`std_swap_ranges<uint8_t>/15`   | 7.72 ns | 7.75 ns | 1.00
`std_swap_ranges<uint8_t>/26`   | 4.96 ns | 4.97 ns | 1.00
`std_swap_ranges<uint8_t>/38`   | 6.71 ns | 6.69 ns | 1.00
`std_swap_ranges<uint8_t>/60`   | 6.30 ns | 6.31 ns | 1.00
`std_swap_ranges<uint8_t>/125`  | 8.02 ns | 7.97 ns | 1.01
`std_swap_ranges<uint8_t>/800`  | 22.5 ns | 23.8 ns | 0.95
`std_swap_ranges<uint8_t>/3000` | 86.2 ns | 89.9 ns | 0.96
`std_swap_ranges<uint8_t>/9000` |  124 ns |  123 ns | 1.01