@StephanTLavavej 2024-10-23
I'm mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.